+++ src/components/pages/dispenser/components/statistics/types.tsx
@@ -3,4 +3,7 @@ import { TDispenserStatus } from "types"
 export type TProps = {
   dispenserStatus: TDispenserStatus
   linksCount: number
+  linksClaimed: number
+  linksAssigned: number
+  downloadReport: () => void
 }