+++ src/data/store/reducers/dispensers/async-actions/download-report.tsx
@@ -1,9 +1,7 @@
 import { Dispatch } from 'redux'
-import { CampaignActions } from '../../campaign/types'
-import { UserActions } from '../../user/types'
 import { RootState } from 'data/store'
-import { CampaignsActions } from '../types'
-import { campaignsApi } from 'data/api'
+import { DispensersActions } from '../types'
+import { dispensersApi } from 'data/api'
 import * as actionsCampaigns from '../actions'
 import {
   downloadLinksAsCSV,
@@ -13,28 +11,28 @@ import {
 import { plausibleApi } from 'data/api'
 
 const downloadReport = (
-  campaignId: string
+  dispenser_id: string
 ) => {
   return async (
-    dispatch: Dispatch<CampaignActions | UserActions | CampaignsActions>,
+    dispatch: Dispatch<DispensersActions>,
     getState: () => RootState
   ) => {
     const { user: { chainId } } = getState()
-    if (!campaignId) { return alertError ('campaignId is not provided') }
+    if (!dispenser_id) { return alertError ('dispenser_id is not provided') }
     try {
     
-      const { data: { links_data } } = await campaignsApi.getReport(campaignId)
+      const { data: { links_data } } = await dispensersApi.getReport(dispenser_id)
       if (links_data.length === 0) {
-        alertError('No data for report. At least one claimed link is needed for report')
+        alertError('No data found for report. At least one claimed link is needed for report')
         return dispatch(actionsCampaigns.setLoading(false))
       }
-      downloadLinksAsCSV(links_data, `REPORT-${campaignId}`)
+      downloadLinksAsCSV(links_data, `REPORT-${dispenser_id}`)
       plausibleApi.invokeEvent({
         eventName: 'download_report',
         data: {
           network: defineNetworkName(chainId),
-          type: 'campaign',
-          campaignId
+          type: 'dispenser',
+          dispenser_id
         }
       })
       dispatch(actionsCampaigns.setLoading(false))