+++ src/data/store/reducers/dispensers/async-actions/index.tsx
@@ -5,13 +5,17 @@ import updateDispenser from './update-dispenser'
 import updateStatus from './update-status'
 import toggleRedirectOn from './toggle-redirect-on'
 import updateRedirectURL from './update-redirect-url'
+import getDispenserStats from './get-dispenser-stats'
+import downloadReport from './download-report'
 
 export {
   addDispenser,
   downloadDispenserQR,
+  downloadReport,
   addLinksToQR,
   updateDispenser,
   updateStatus,
   toggleRedirectOn,
-  updateRedirectURL
+  updateRedirectURL,
+  getDispenserStats
 }