+++ src/data/store/reducers/dispensers/async-actions/update-redirect-url.tsx
@@ -5,10 +5,13 @@ import { RootState } from 'data/store'
 import { dispensersApi } from 'data/api'
 import { alertError } from 'helpers'
 import { plausibleApi } from 'data/api'
+import { decrypt, encrypt } from 'lib/crypto'
+import { ethers } from 'ethers'
 
 type TUpdateRedirectURLArgs = {
   dispenser_id: string
   redirect_url: string
+  encrypted_multiscan_qr_enc_code: string
   successCallback: () => void
   errorCallback: () => void
 }
@@ -16,6 +19,7 @@ type TUpdateRedirectURLArgs = {
 const updateRedirectURL = ({
   dispenser_id,
   redirect_url,
+  encrypted_multiscan_qr_enc_code,
   successCallback,
   errorCallback
 }: TUpdateRedirectURLArgs) => {
@@ -23,14 +27,17 @@ const updateRedirectURL = ({
     dispatch: Dispatch<DispensersActions>,
     getState: () => RootState
   ) => {
-    const { user: { address }, dispensers: { dispensers } } = getState()
+    const { user: { address, dashboardKey }, dispensers: { dispensers } } = getState()
     dispatch(actionsDispensers.setLoading(true))
     try {
-      const { data } : { data: { success: boolean } } = await dispensersApi.updateRedirectUrl({ dispenser_id, redirect_url })
+      const multiscanQREncCode = decrypt(encrypted_multiscan_qr_enc_code, dashboardKey as string)
+      const linkKey = ethers.utils.id(multiscanQREncCode)
+      const redirectLinkEncrypted = encrypt(redirect_url, linkKey.replace('0x', ''))
+      const { data } : { data: { success: boolean } } = await dispensersApi.updateRedirectUrl({ dispenser_id, redirect_url: redirectLinkEncrypted })
       if (data.success) {
         const dispensersUpdated = dispensers.map(item => {
           if (item.dispenser_id === dispenser_id) { 
-            return { ...item, redirect_url }
+            return { ...item, redirect_url: redirectLinkEncrypted }
           }
           return item
         })