+++ src/types/dispenser.tsx
@@ -22,8 +22,10 @@ export type TDispenser = {
   title: string
   encrypted_multiscan_qr_enc_code: string
   active?: boolean
-  redirect_on: boolean
-  redirect_url: string | null
+  redirect_on?: boolean
+  redirect_url?: string | null
+  links_claimed?: number
+  links_assigned?: number
 }
 
 export type TDispenserUpdateData = {